requirements.txt
# Email to Email AI Agent Dependencies
# Install with: py -3 -m pip install -r requirements.txt

# YAML configuration support
PyYAML>=6.0

# Google Gemini for intelligent replies (FREE!)
google-generativeai>=0.3.0